Move pictures loader in a dedicated file
